request: move cache checker into ut
