+++ request/testdata/dummy_nonexistent_kubeconfig.yaml
@@ -1,0 +1,21 @@
+apiVersion: v1
+clusters:
+- cluster:
+    server: https://unit-test.kperf.io
+  name: unit-test.kperf.io
+contexts:
+- context:
+    cluster: unit-test.kperf.io
+    user: testing@unit-test.kperf.io
+  name: testing@unit-test.kperf.io
+current-context: testing@unit-test.kperf.io
+kind: Config
+preferences: {}
+users:
+- name: testing@unit-test.kperf.io
+  user:
+    exec:
+      apiVersion: client.authentication.k8s.io/v1beta1
+      command: unknown
+      interactiveMode: IfAvailable
+      provideClusterInfo: false
